perf: improve performance of error creation
The stack of global errors is omitted. Previously the check, if a error is global, was performed on error creation, which required us to generate the frames and check if the stack trace contained a function call to `runtime.doInit`. This is a kind of expensive operation and therefore it is now performed when the error is formatted. While formatting the error the frames are generated anyway and thus the check is almost free.